refactor: common-module passport 폴더 구조 변경
