+++ src/backend/common-module/build.gradle
@@ -21,6 +21,7 @@ repositories {
 dependencies {
 	// Spring
 	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
 	// Validation
 	implementation 'jakarta.validation:jakarta.validation-api:3.0.2'
