+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/global/response/code/success/ReasonDTO.java
@@ -1,4 +1,4 @@
-package com.bbebig.commonmodule.response.code.success;
+package com.bbebig.commonmodule.global.response.code.success;
 
 import lombok.Builder;
 import lombok.Getter;
